refactor(cost): extract effective FX enrichment
