+++ src/services/portfolio_transaction_processing_service/app/application/cost_basis_processing/__init__.py
@@ -1,5 +1,6 @@
-"""Expose application policy for preparing booked transactions for cost processing."""
+"""Expose application policy for canonical cost-basis processing."""
 
+from .fx_enrichment import FxRateNotFoundError, enrich_cost_basis_transactions_with_fx
 from .preparation import (
     CostProcessingRoute,
     InstrumentReferenceUnavailableError,
@@ -9,7 +10,9 @@ from .preparation import (
 
 __all__ = [
     "CostProcessingRoute",
+    "FxRateNotFoundError",
     "InstrumentReferenceUnavailableError",
     "PreparedCostTransaction",
+    "enrich_cost_basis_transactions_with_fx",
     "prepare_cost_transaction",
 ]
